+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 🐛 *Bug Fixes*
 
+* Properly handle QE deprecated and connection timeout errors when using `WorkflowRun.by_id()` without `config` parameter passed
+
 💅 *Improvements*
 
 🥷 *Internal*
@@ -35,6 +37,7 @@
 🐛 *Bug Fixes*
 
 * Workaround for Ray cluster not starting because of a missing dependency, `async_timeout`.
+* Fix warnings when used with `aiohttp>3.9`.
 
 💅 *Improvements*